+++ test/framework/src/main/java/org/elasticsearch/test/rest/section/MatchAssertion.java
@@ -18,15 +18,21 @@
  */
 package org.elasticsearch.test.rest.section;
 
+import org.elasticsearch.common.Nullable;
 import org.elasticsearch.common.logging.ESLogger;
 import org.elasticsearch.common.logging.Loggers;
 
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.Objects;
+import java.util.TreeMap;
 import java.util.regex.Pattern;
 
 import static org.elasticsearch.test.hamcrest.RegexMatcher.matches;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.instanceOf;
-import static org.hamcrest.Matchers.notNullValue;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
 
 /**
@@ -45,7 +51,6 @@ public class MatchAssertion extends Assertion {
 
     @Override
     protected void doAssert(Object actualValue, Object expectedValue) {
-
         //if the value is wrapped into / it is a regexp (e.g. /s+d+/)
         if (expectedValue instanceof String) {
             String expValue = ((String) expectedValue).trim();
@@ -60,20 +65,131 @@ public class MatchAssertion extends Assertion {
             }
         }
 
-        assertThat(errorMessage(), actualValue, notNullValue());
+        assertNotNull("field [" + getField() + "] is null", actualValue);
         logger.trace("assert that [{}] matches [{}] (field [{}])", actualValue, expectedValue, getField());
-        if (!actualValue.getClass().equals(safeClass(expectedValue))) {
+        if (actualValue.getClass().equals(safeClass(expectedValue)) == false) {
             if (actualValue instanceof Number && expectedValue instanceof Number) {
                 //Double 1.0 is equal to Integer 1
-                assertThat(errorMessage(), ((Number) actualValue).doubleValue(), equalTo(((Number) expectedValue).doubleValue()));
+                assertThat("field [" + getField() + "] doesn't match the expected value",
+                        ((Number) actualValue).doubleValue(), equalTo(((Number) expectedValue).doubleValue()));
                 return;
             }
         }
 
-        assertThat(errorMessage(), actualValue, equalTo(expectedValue));
+        if (expectedValue.equals(actualValue) == false) {
+            FailureMessage message = new FailureMessage(getField());
+            message.compare(getField(), actualValue, expectedValue);
+            throw new AssertionError(message.message);
+        }
     }
 
-    private String errorMessage() {
-        return "field [" + getField() + "] doesn't match the expected value";
+    private static class FailureMessage {
+        private final StringBuilder message;
+        private int indent = 0;
+
+        private FailureMessage(String field) {
+            this.message = new StringBuilder(field + " didn't match the expected value:\n");
+        }
+
+        private void compareMaps(Map<String, Object> actual, Map<String, Object> expected) {
+            actual = new TreeMap<>(actual);
+            expected = new TreeMap<>(expected);
+            for (Map.Entry<String, Object> expectedEntry : expected.entrySet()) {
+                compare(expectedEntry.getKey(), actual.remove(expectedEntry.getKey()), expectedEntry.getValue());
+            }
+            for (Map.Entry<String, Object> unmatchedEntry : actual.entrySet()) {
+                field(unmatchedEntry.getKey(), "unexpected but found [" + unmatchedEntry.getValue() + "]");
+            }
+        }
+
+        private void compareLists(List<Object> actual, List<Object> expected) {
+            int i = 0;
+            while (i < actual.size() && i < expected.size()) {
+                compare(Integer.toString(i), actual.get(i), expected.get(i));
+                i++;
+            }
+            if (actual.size() == expected.size()) {
+                return;
+            }
+            indent();
+            if (actual.size() < expected.size()) {
+                message.append("expected [").append(expected.size() - i).append("] more entries\n");
+                return;
+            }
+            message.append("received [").append(actual.size() - i).append("] more entries than expected\n");
+        }
+
+        private void compare(String field, @Nullable Object actual, Object expected) {
+            if (expected instanceof Map) {
+                if (actual == null) {
+                    field(field, "expected map but not found");
+                    return;
+                }
+                if (false == actual instanceof Map) {
+                    field(field, "expected map but found [" + actual + "]");
+                    return;
+                }
+                @SuppressWarnings("unchecked")
+                Map<String, Object> expectedMap = (Map<String, Object>) expected;
+                @SuppressWarnings("unchecked")
+                Map<String, Object> actualMap = (Map<String, Object>) actual;
+                if (expectedMap.isEmpty() && actualMap.isEmpty()) {
+                    field(field, "same [empty map]");
+                    return;
+                }
+                field(field, null);
+                indent += 1;
+                compareMaps(actualMap, expectedMap);
+                indent -= 1;
+                return;
+            }
+            if (expected instanceof List) {
+                if (actual == null) {
+                    field(field, "expected list but not found");
+                    return;
+                }
+                if (false == actual instanceof List) {
+                    field(field, "expected list but found [" + actual + "]");
+                    return;
+                }
+                @SuppressWarnings("unchecked")
+                List<Object> expectedList = (List<Object>) expected;
+                @SuppressWarnings("unchecked")
+                List<Object> actualList = (List<Object>) actual;
+                if (expectedList.isEmpty() && actualList.isEmpty()) {
+                    field(field, "same [empty list]");
+                    return;
+                }
+                field(field, null);
+                indent += 1;
+                compareLists(actualList, expectedList);
+                indent -= 1;
+                return;
+            }
+            if (actual == null) {
+                field(field, "expected [" + expected + "] but not found");
+                return;
+            }
+            if (Objects.equals(expected, actual)) {
+                field(field, "same [" + expected + "]");
+                return;
+            }
+            field(field, "expected [" + expected + "] but was [" + actual + "]");
+        }
+
+        private void indent() {
+            for (int i = 0; i < indent; i++) {
+                message.append("  ");
+            }
+        }
+
+        private void field(Object name, String info) {
+            indent();
+            message.append(String.format(Locale.ROOT, "%30s: ", name));
+            if (info != null) {
+                message.append(info);
+            }
+            message.append('\n');
+        }
     }
 }
